Add option to update kata shim
Following are the key files

- kata-shim-ds.yaml: This is the DS spec to deploy custom Kata shim for
  CoCo. The shim will be copied to
  /opt/kata/bin/containerd-shim-kata-v2-coco
- mc-40-kata-config.yaml: This is the MachineConfig which is used to
  update the CRIO config to align with the custom shim

The source crio config files for the MachineConfig are under `local:` and
the butane config is available in butane-crio.yaml

Also note that the MC is indexed with digit `60`. This ensures that this
specific MC will be applied after the original one for peer-pods which
is indexed with digit `40`. This is needed to avoid MCO errors.

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>